Pin remaining stale float indices to integer in tests
Two more tests built indices via default-float NDArrays and relied on
the old auto-cast that the branch's strict integer-index check removed:
- test_ndarray_slice: advanced index arrays -> int64
- test_random multinomial generator helper: scatter_nd index -> int64

+++ tests/python/unittest/test_ndarray.py
@@ -546,7 +546,8 @@ def test_ndarray_slice():
 
     a = mx.nd.array([[0, 1], [2, 3]])
     assert (a[[1, 1, 0], [0, 1, 0]].asnumpy() == [2, 3, 0]).all()
-    assert (a[mx.nd.array([1, 1, 0]), mx.nd.array([0, 1, 0])].asnumpy() == [2, 3, 0]).all()
+    assert (a[mx.nd.array([1, 1, 0], dtype='int64'),
+              mx.nd.array([0, 1, 0], dtype='int64')].asnumpy() == [2, 3, 0]).all()
 
     shape = (4, 4)
     A = mx.nd.random.uniform(shape=shape)
